feat : festival 타입 분리
