+++ app/lib/types/index.ts
@@ -1,1 +1,2 @@
 export * from './types';
+export { type FestivalImage, type FestivalInfo } from './festival';
